+++ app/lib/main.dart
@@ -1,6 +1,5 @@
 import 'dart:io';
 
-import 'package:bitsdojo_window/bitsdojo_window.dart';
 import 'package:flutter/material.dart';
 import 'package:for_net_ui/mobile_app.dart';
 import 'package:for_net_ui/pc_app.dart';
@@ -14,16 +13,4 @@ void main() {
   }
 }
 
-void pcRun() {
-  runApp(const PCApp());
-  doWhenWindowReady(() {
-    final win = appWindow;
-    const initialSize = Size(400, 160);
-    win.minSize = initialSize;
-    win.size = initialSize;
-    win.alignment = Alignment.center;
-    //win.show();
-    //win.hide();
-  });
-}
