+++ app/lib/pc_app.dart
@@ -8,8 +8,8 @@ import 'package:bitsdojo_window/bitsdojo_window.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:for_net_ui/native/ffi.dart';
-import 'package:for_net_ui/page/login_page.dart';
-import 'package:for_net_ui/page/welcome_page.dart';
+import 'package:for_net_ui/page/pc/login_page.dart';
+import 'package:for_net_ui/page/pc/welcome_page.dart';
 import 'package:launch_at_startup/launch_at_startup.dart';
 import 'package:system_tray/system_tray.dart';
 import 'package:package_info_plus/package_info_plus.dart';
@@ -29,6 +29,19 @@ class PCApp extends StatefulWidget {
 
 bool initSuccess = false;
 
+
+void pcRun() {
+  runApp(const PCApp());
+  doWhenWindowReady(() {
+    final win = appWindow;
+    const initialSize = Size(400, 160);
+    win.minSize = initialSize;
+    win.size = initialSize;
+    win.alignment = Alignment.center;
+    //win.show();
+    //win.hide();
+  });
+}
 class _PCAppState extends State<PCApp> {
   final AppWindow _appWindow = AppWindow();
   final SystemTray _systemTray = SystemTray();
